backend/requirements.txt
django
dj-database-url
whitenoise
djangorestframework
djangorestframework-simplejwt
python-decouple
Pillow
drf-spectacular
stripe
django-cors-headers
psycopg2-binary
dj-database-url
redis
celery
channels
channels-redis
daphne